app2/requirements.txt
fastapi
uvicorn
psycopg2-binary
pgvector
redis
sentence-transformers
pydantic
python-dotenv
groq
requests
beautifulsoup4
paho-mqtt
numpy
torch
transformers
paho-mqtt
pillow
opencv-python
pytesseract
numpy
